JESS: move globally used roundMulti function from #2 to a beter place (dvb.h/dvb.cpp)
This enables the prototype to be checked by the compiler (it's
seen both when compiling the function and when calling it from
another compilation unit and that's how it's supposed to work,
don't add isolated prototypes or external references inside a
source file where it's called, it's useless then).

Remove useless "volatile" qualifier.