[Chanel selection context menu] Neat ordering of entries. (#498)
I do not want to carch it up... But it are keys for move, copy etc.   and we prefer to keep these as they were choosen that way in the past... So I will accept this when the order is changed and e.g. 6 stays on move.... (In basics I like this recommeation....)